Move fields from dataset to variable level
